+++ python/tests/engine/test_python.py
@@ -3503,7 +3503,7 @@ class TestPython:
         # Arrange
         mocker.patch("hsfs.core.kafka_engine.get_kafka_config", return_value={})
         mocker.patch("hsfs.feature_group.FeatureGroup._get_encoded_avro_schema")
-        mock_save_ingestion_run = mocker.patch("hsfs.core.feature_group_api.FeatureGroupApi.save_ingestion_run")
+        mock_save_ingestion_run = mocker.patch("hsfs.core.ingestion_run_api.IngestionRunApi.save_ingestion_run")
         mocker.patch("hsfs.core.kafka_engine.get_encoder_func")
         mocker.patch("hsfs.core.kafka_engine.encode_complex_features")
         mock_python_engine_kafka_produce = mocker.patch(
@@ -3563,7 +3563,7 @@ class TestPython:
         # Arrange
         mocker.patch("hsfs.core.kafka_engine.get_kafka_config", return_value={})
         mocker.patch("hsfs.feature_group.FeatureGroup._get_encoded_avro_schema")
-        mock_save_ingestion_run = mocker.patch("hsfs.core.feature_group_api.FeatureGroupApi.save_ingestion_run")
+        mock_save_ingestion_run = mocker.patch("hsfs.core.ingestion_run_api.IngestionRunApi.save_ingestion_run")
         mocker.patch("hsfs.core.kafka_engine.get_encoder_func")
         mocker.patch("hsfs.core.kafka_engine.encode_complex_features")
         mock_python_engine_kafka_produce = mocker.patch(
@@ -3623,7 +3623,7 @@ class TestPython:
         # Arrange
         mocker.patch("hsfs.core.kafka_engine.get_kafka_config", return_value={})
         mocker.patch("hsfs.feature_group.FeatureGroup._get_encoded_avro_schema")
-        mock_save_ingestion_run = mocker.patch("hsfs.core.feature_group_api.FeatureGroupApi.save_ingestion_run")
+        mock_save_ingestion_run = mocker.patch("hsfs.core.ingestion_run_api.IngestionRunApi.save_ingestion_run")
         mocker.patch("hsfs.core.kafka_engine.get_encoder_func")
         mocker.patch("hsfs.core.kafka_engine.encode_complex_features")
         mock_python_engine_kafka_produce = mocker.patch(
@@ -3682,7 +3682,7 @@ class TestPython:
         # Arrange
         mocker.patch("hsfs.core.kafka_engine.get_kafka_config", return_value={})
         mocker.patch("hsfs.feature_group.FeatureGroup._get_encoded_avro_schema")
-        mock_save_ingestion_run = mocker.patch("hsfs.core.feature_group_api.FeatureGroupApi.save_ingestion_run")
+        mock_save_ingestion_run = mocker.patch("hsfs.core.ingestion_run_api.IngestionRunApi.save_ingestion_run")
         mocker.patch("hsfs.core.kafka_engine.get_encoder_func")
         mocker.patch("hsfs.core.kafka_engine.encode_complex_features")
         mock_python_engine_kafka_produce = mocker.patch(
@@ -3769,7 +3769,7 @@ class TestPython:
             "hsfs.core.kafka_engine.kafka_get_offsets",
             side_effect=["tests_offsets1", "tests_offsets2"],
         )
-        mock_fg_api = mocker.patch("hsfs.core.feature_group_api.FeatureGroupApi")
+        mock_save_ingestion_run = mocker.patch("hsfs.core.ingestion_run_api.IngestionRunApi.save_ingestion_run")
         mocker.patch(
             "hsfs.core.job_api.JobApi.last_execution",
             return_value=[],
@@ -3807,8 +3807,8 @@ class TestPython:
         )
 
         # Assert
-        assert mock_fg_api.return_value.save_ingestion_run.call_count == 1
-        args, _ = mock_fg_api.return_value.save_ingestion_run.call_args
+        assert mock_save_ingestion_run.call_count == 1
+        args, _ = mock_save_ingestion_run.call_args
         assert args[0] == fg
         assert args[1].starting_offsets == "tests_offsets1"
         assert args[1].ending_offsets == "tests_offsets2"
