+++ python/tests/test_feature_group_writer.py
@@ -21,7 +21,7 @@ from hsfs.engine import python
 class TestFeatureGroupWriter:
     def test_fg_writer_context_manager(self, mocker, dataframe_fixture_basic):
         mock_insert = mocker.patch("hsfs.feature_group.FeatureGroup.insert")
-        mock_save_ingestion_run = mocker.patch("hsfs.core.feature_group_api.FeatureGroupApi.save_ingestion_run")
+        mock_save_ingestion_run = mocker.patch("hsfs.core.ingestion_run_api.IngestionRunApi.save_ingestion_run")
 
         fg = feature_group.FeatureGroup(
             name="test",
@@ -53,7 +53,7 @@ class TestFeatureGroupWriter:
         mocker.patch("hsfs.engine.get_instance", return_value=engine)
         mocker.patch("hopsworks_common.client.get_instance")
         mocker.patch("hsfs.core.kafka_engine.kafka_get_offsets", return_value="")
-        mock_save_ingestion_run = mocker.patch("hsfs.core.feature_group_api.FeatureGroupApi.save_ingestion_run")
+        mock_save_ingestion_run = mocker.patch("hsfs.core.ingestion_run_api.IngestionRunApi.save_ingestion_run")
         producer, feature_writers, writer_m = (
             mocker.MagicMock(),
             mocker.MagicMock(),
@@ -114,7 +114,7 @@ class TestFeatureGroupWriter:
         mocker.patch("hsfs.engine.get_instance", return_value=engine)
         mocker.patch("hopsworks_common.client.get_instance")
         mocker.patch("hsfs.core.kafka_engine.kafka_get_offsets", return_value="")
-        mock_save_ingestion_run = mocker.patch("hsfs.core.feature_group_api.FeatureGroupApi.save_ingestion_run")
+        mock_save_ingestion_run = mocker.patch("hsfs.core.ingestion_run_api.IngestionRunApi.save_ingestion_run")
         producer, feature_writers, writer_m = (
             mocker.MagicMock(),
             mocker.MagicMock(),
